+++ ta4j-core/src/test/java/org/ta4j/core/TestUtils.java
@@ -82,6 +82,9 @@ public class TestUtils {
      *                        {@code int} representation
      */
     public static void assertNumEquals(int expected, Num actual) {
+        if (actual.isNaN()) {
+            throw new AssertionError("Expected: " + expected + " Actual: " + actual);
+        }
         assertEquals(actual.numOf(expected), actual);
     }