+++ ta4j-core/src/test/java/org/ta4j/core/indicators/aroon/AroonOscillatorIndicatorTest.java
@@ -21,7 +21,7 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package org.ta4j.core.indicators;
+package org.ta4j.core.indicators.aroon;
 
 import static org.junit.Assert.assertNotNull;
 import static org.ta4j.core.TestUtils.assertNumEquals;
@@ -36,6 +36,7 @@ import org.junit.Before;
 import org.junit.Test;
 import org.ta4j.core.BarSeries;
 import org.ta4j.core.BaseBarSeries;
+import org.ta4j.core.indicators.aroon.AroonOscillatorIndicator;
 
 public class AroonOscillatorIndicatorTest {
     private BarSeries data;