+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 ## 0.15 (unreleased)
 
 ### Breaking
+- **NumberOfConsecutiveWinningPositions** renamed to **`NumberOfConsecutivePositions`**
 - **DifferencePercentage** renamed to **`DifferencePercentageIndicator`**
 - **BuyAndHoldCriterion** renamed to **`EnterAndHoldCriterion`**
 - **DXIndicator** moved to adx-package
@@ -10,6 +11,7 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 - **MinusDMIndicator** moved to adx-package
 - `analysis/criterion`-package moved to root
 - `cost`-package moved to `analysis/cost`-package
+- **AroonXXX** indicators moved to aroon package
 
 ### Fixed
 - **LosingPositionsRatioCriterion** correct betterThan
@@ -17,6 +19,8 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 - **Fixed** **`ChaikinOscillatorIndicatorTest`**
 - **DecimalNum#remainder()** adds NaN-check 
 - **Fixed** **ParabolicSarIndicatorTest** fixed openPrice always 0 and highPrice lower than lowPrice
+- **UlcerIndexIndicator** using the max price of current period instead of the highest value of last n bars
+- **DurationBarAggregator** fixed aggregation of bars with gaps
 
 ### Changed
 - **KeltnerChannelMiddleIndicator** changed superclass to AbstractIndicator; add GetBarCount() and toString()
@@ -33,6 +37,8 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 - **PriceIndicator** removed
 
 ### Added
+- **NumericIndicator** new class providing a fluent and lightweight api for indicator creation
+- **AroonFacade**, **BollingerBandFacade**, **KeltnerChannelFacade** new classes providing a facade for indicator groups by using lightweight `NumericIndicators`
 - **AbstractEMAIndicator** added getBarCount() to support future enhancements 
 - **ATRIndicator** "uncached" by changing superclass to AbstractIndicator; added constructor to accept TRIndicator and getter for same; added toString(); added getBarCount() to support future enhancements
 - :tada: **Enhancement** added possibility to use CostModels when backtesting with the BacktestExecutor
@@ -53,6 +59,7 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 - **Enhancement** prevent Object[] allocation for varargs log.trace and log.debug calls by wrapping them in `if` blocks
 - :tada: **Enhancement** added **`FixedTransactionCostModel`**
 - :tada: **Enhancement** added **`WaitForRule#noWaitIfNoTradeExists`** no need to wait if no trade exists
+- :tada: **Enhancement** added **`AnalysisCriterion.PositionFilter`** to handle both sides within one Criterion.
 
 ## 0.14 (released April 25, 2021)